% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_types.R
\name{assign_types}
\alias{assign_types}
\title{Set the types of the columns of a dataframe}
\usage{
assign_types(data, types)
}
\arguments{
\item{data}{dataframe}

\item{types}{containing the types to set to the columns
'types': {'column_name': {'type': 'numeric' | 'factor'},
'column_name': {
'type': 'factor',
'levels': \link{'a', 'b', 'c'},
'ref': 'a'
}
}}
}
\value{
dataframe with the specified types
}
\description{
Set the types of the columns of a dataframe
}
