% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize}
\alias{factorize}
\title{Convert character variables to factors}
\usage{
factorize(data)
}
\arguments{
\item{data}{data.frame}
}
\value{
data.frame
}
\description{
Convert character variables to factors
}
